backup/arbeidskrav_2/requirements.txt
mysql-connector-python>=8.0
python-dotenv>=0.21.0